Flatten external sorters and test core sorters
