+++ conftest.py
@@ -9,7 +9,8 @@ ON_GITHUB = bool(os.getenv('GITHUB_ACTIONS'))
 
 # define marks
 mark_names = ["core", "extractors", "preprocessing", "postprocessing",
-              "sorters_external", "sorters_internal", "qualitymetrics", "comparison", "curation",
+              "sorters_external", "sorters_internal", "sorters",
+              "qualitymetrics", "comparison", "curation",
               "widgets", "exporters", "sortingcomponents"]
 
 
@@ -37,17 +38,19 @@ def pytest_collection_modifyitems(config, items):
 
     for item in items:
         rel_path = Path(item.fspath).relative_to(rootdir)
-
-        if "sorters" not in rel_path.name:
+        if "sorters" not in str(rel_path):
             for mark_name in mark_names:
                 if f"/{mark_name}/" in str(rel_path):
                     mark = getattr(pytest.mark, mark_name)
                     item.add_marker(mark)
         else:
-            if "internal" in rel_path.name:
+            if "/internal/" in str(rel_path):
                 item.add_marker("sorters_internal")
-            elif "external" in rel_path.name:
+            elif "/external/" in str(rel_path):
                 item.add_marker("sorters_external")
+            else:
+                print(rel_path)
+                item.add_marker("sorters")
 
 
 def pytest_sessionfinish(session, exitstatus):
