+++ spikeinterface/sorters/external/combinato.py
@@ -4,9 +4,9 @@ from typing import Union
 import sys
 import json
 
-from ...utils import ShellScript
+from ..utils import ShellScript
 from spikeinterface.core import write_to_h5_dataset_format
-from ...basesorter import BaseSorter
+from ..basesorter import BaseSorter
 from spikeinterface.extractors import CombinatoSortingExtractor
 from spikeinterface.preprocessing import ScaleRecording
 
@@ -153,7 +153,7 @@ class CombinatoSorter(BaseSorter):
         tmpdir = sorter_output_folder
 
         if verbose:
-            print(f'Running combinato in {tmpdir}...')
+            print(f'Running combinato in {tmpdir}..')
 
         with open(tmpdir / "local_options.py", "w") as outFile:
             outFile.writelines("options = {}".format(p))
