--- spikeinterface/sorters/external/combinato/__init__.py
@@ -1,1 +1,0 @@
-from .combinato import CombinatoSorter
