+++ spikeinterface/sorters/external/hdsort.py
@@ -8,8 +8,8 @@ import numpy as np
 import scipy.io
 
 from spikeinterface.core.core_tools import write_to_h5_dataset_format
-from ...basesorter import BaseSorter
-from ...utils import ShellScript
+from ..basesorter import BaseSorter
+from ..utils import ShellScript
 
 # from spikeinterface.extractors import MaxOneRecordingExtractor
 from spikeinterface.extractors import HDSortSortingExtractor
@@ -151,7 +151,7 @@ class HDSortSorter(BaseSorter):
         }
         P['featureExtraction'] = {'nDims': float(params['n_pc_dims'])}  # 6
         P['clustering'] = {
-            'maxSpikes': 50000.0,  # dont align spikes you dont cluster...
+            'maxSpikes': 50000.0,  # dont align spikes you dont cluster..
             'meanShiftBandWidthFactor': 1.8
             # 'meanShiftBandWidth': sqrt(1.8*6)  # todo: check this!
         }
