--- spikeinterface/sorters/external/hdsort/__init__.py
@@ -1,1 +1,0 @@
-from .hdsort import HDSortSorter
