--- spikeinterface/sorters/external/herdingspikes/__init__.py
@@ -1,1 +1,0 @@
-from .herdingspikes import HerdingspikesSorter
