--- spikeinterface/sorters/external/ironclust/__init__.py
@@ -1,1 +1,0 @@
-from .ironclust import IronClustSorter
