+++ spikeinterface/sorters/external/kilosort.py
@@ -3,9 +3,9 @@ import os
 from typing import Union
 import numpy as np
 
-from ...basesorter import BaseSorter
-from ..kilosortbase import KilosortBase
-from ...utils import get_git_commit
+from ..basesorter import BaseSorter
+from .kilosortbase import KilosortBase
+from ..utils import get_git_commit
 
 
 def check_if_installed(kilosort_path: Union[str, None]):
