--- spikeinterface/sorters/external/kilosort/__init__.py
@@ -1,1 +1,0 @@
-from .kilosort import KilosortSorter
