--- spikeinterface/sorters/external/kilosort2/__init__.py
@@ -1,1 +1,0 @@
-from .kilosort2 import Kilosort2Sorter
