--- spikeinterface/sorters/external/kilosort2_5/__init__.py
@@ -1,1 +1,0 @@
-from .kilosort2_5 import Kilosort2_5Sorter
