--- spikeinterface/sorters/external/kilosort3/__init__.py
@@ -1,1 +1,0 @@
-from .kilosort3 import Kilosort3Sorter
