+++ spikeinterface/sorters/external/klusta.py
@@ -3,8 +3,8 @@ from pathlib import Path
 import sys
 import shutil
 
-from ...basesorter import BaseSorter, get_job_kwargs
-from ...utils import ShellScript
+from ..basesorter import BaseSorter, get_job_kwargs
+from ..utils import ShellScript
 
 from probeinterface import write_prb
 
@@ -113,7 +113,7 @@ class KlustaSorter(BaseSorter):
             detect_sign = 'both'
 
         # set up klusta config file
-        with (source_dir / 'config_default.prm').open('r') as f:
+        with (source_dir / 'klusta_config_default.prm').open('r') as f:
             klusta_config = f.readlines()
 
         # Note: should use format with dict approach here
