--- spikeinterface/sorters/external/klusta/__init__.py
@@ -1,1 +1,0 @@
-from .klusta import KlustaSorter
