+++ spikeinterface/sorters/external/mountainsort4.py
@@ -4,7 +4,7 @@ from packaging.version import parse
 
 from spikeinterface.preprocessing import bandpass_filter, whiten
 
-from ...basesorter import BaseSorter
+from ..basesorter import BaseSorter
 from spikeinterface.core.old_api_utils import NewToOldRecording
 from spikeinterface.core import load_extractor
