--- spikeinterface/sorters/external/mountainsort4/__init__.py
@@ -1,1 +1,0 @@
-from .mountainsort4 import Mountainsort4Sorter
