--- spikeinterface/sorters/external/pykilosort/__init__.py
@@ -1,1 +1,0 @@
-from .pykilosort import PyKilosortSorter
