--- spikeinterface/sorters/external/spyking_circus/__init__.py
@@ -1,1 +1,0 @@
-from .spyking_circus import SpykingcircusSorter
