--- spikeinterface/sorters/external/tridesclous/__init__.py
@@ -1,1 +1,0 @@
-from .tridesclous import TridesclousSorter
