--- spikeinterface/sorters/external/waveclus/__init__.py
@@ -1,2 +1,0 @@
-from .waveclus import WaveClusSorter
-from .waveclus_snippets import WaveClusSnippetsSorter
