--- spikeinterface/sorters/external/yass/__init__.py
@@ -1,1 +1,0 @@
-from .yass import YassSorter
